civic-complaint-system/app/utils/geocode.py
import requests


def reverse_geocode(lat: float, lng: float) -> str:
    url = "https://nominatim.openstreetmap.org/reverse"
    params = {
        "lat": lat,
        "lon": lng,
        "format": "json"
    }
    headers = {
        "User-Agent": "civic-complaint-system"
    }

    response = requests.get(url, params=params, headers=headers, timeout=10)

    if response.status_code != 200:
        return "Unknown location"

    data = response.json()
    return data.get("display_name", "Unknown location")
